chore: split the permission forms UI code so it's possible to add a new tab [PT-187612692]
