+++ rails/react-components/src/library/components/permission-forms-v2/manage-forms-tab.scss
@@ -1,0 +1,52 @@
+@import "../../../shared/styles/variables/variables";
+
+.manageFormsTabContent {
+  .controlsArea {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 20px;
+
+    label {
+      text-transform: none;
+      font-weight: 300;
+      margin-right: 10px;
+      display: flex;
+      align-items: center;
+    }
+
+    div {
+      display: flex;
+    }
+
+    .leftSide {
+      width: 50%;
+      justify-content: flex-start;
+      > div {
+        align-items: center;
+        margin-right:15px;
+      }
+      select {
+        width: 400px;
+      }
+    }
+
+    .rightSide {
+      justify-content: flex-end;
+    }
+  }
+  .permissionFormsTable {
+    margin-top:20px;
+    width: 100%;
+
+    tr {
+      // every other row is background #f9f9f9
+      &:nth-child(odd) {
+        background-color: #f9f9f9;
+      }
+    }
+    th {
+      text-align: left;
+      background-color: white;
+    }
+  }
+}
