+++ rails/react-components/src/library/components/permission-forms-v2/permission-form-types.ts
@@ -15,4 +15,4 @@ export interface IProject {
   name: string;
 }
 
-export type PermissionsTab = "projectsTab" | "studentsTab";
+export type PermissionsTab = "manageFormsTab" | "studentsTab";
