--- rails/react-components/src/library/components/permission-forms-v2/style.scss
@@ -1,87 +1,0 @@
-@import "../../../shared/styles/variables/variables";
-
-
-.permissionForms {
-  h2, h3 {
-    color: $col-link;
-    font-family: museo-sans, verdana, helvetica, arial, sans-serif;
-  }
-
-  h3 {
-    padding-top:20px;
-    margin-bottom: 20px;
-    border-top: 1px solid #aaa;
-  }
-
-  .tabArea {
-    .tabs {
-      display:flex;
-      justify-content: left;
-      margin-bottom: 20px;
-
-      a {
-        margin-right: 20px;
-        color: $col-blue;
-        &:hover {
-          color: lighten($col-blue, 10%);
-        }
-      }
-
-      .activeTab {
-        color: $col-link;
-        font-weight: 500;
-        pointer-events: none;
-      }
-    }
-  }
-  .projectPermissionsTabContent {
-    .controlsArea {
-      display: flex;
-      justify-content: space-between;
-      margin-bottom: 20px;
-
-      label {
-        text-transform: none;
-        font-weight: 300;
-        margin-right: 10px;
-        display: flex;
-        align-items: center;
-      }
-
-      div {
-        display: flex;
-      }
-
-      .leftSide {
-        width: 50%;
-        justify-content: flex-start;
-        > div {
-          align-items: center;
-          margin-right:15px;
-        }
-        select {
-          width: 400px;
-        }
-      }
-
-      .rightSide {
-        justify-content: flex-end;
-      }
-    }
-    .permissionFormsTable {
-      margin-top:20px;
-      width: 100%;
-
-      tr {
-        // every other row is background #f9f9f9
-        &:nth-child(odd) {
-          background-color: #f9f9f9;
-        }
-      }
-      th {
-        text-align: left;
-        background-color: white;
-      }
-    }
-  }
-}
